requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
PyMySQL==1.1.1
pytest==8.3.4
PyYAML==6.0.2
requests==2.32.3
tomli==2.2.1
urllib3==2.2.3
